The project is divided into files, default and extra ignore configuration add
