--- .unimport.cfg
@@ -1,3 +1,0 @@
-build
-dist
-.git
