+++ unimport/.unimport.cfg
@@ -1,0 +1,32 @@
+[ignore]
+.*(.git)
+.*(.github)
+.*(build)
+.*(__pycache__)
+.*(develop-eggs)
+.*(dist)
+.*(downloads)
+.*(eggs)
+.*(lib)
+.*(lib64)
+.*(parts)
+.*(sdist)
+.*(var)
+.*(wheels)
+.*(.egg-info)
+.*(MANIFEST)
+.*(htmlcov)
+.*(.tox)
+.*(.hypothesis)
+.*(.pytest_cache)
+.*(local_settings.py)
+.*(instance)
+.*(docs)
+.*(target)
+.*(celerybeat-schedule)
+.*(.sage.py)
+.*(.venv)
+.*(env)
+.*(venv)
+.*(site)
+.*(.mypy_cache)
